Add primary_ip_family to ssconf
Since this parameter will be used on all daemon startups, it needs to be
available on all nodes. This is achieved by querying it via ssconf. This
patch additionally adds a getter method to readily retrieve the primary
ip family from a ConfigWriter object.

This patch also disables the 'R0904: Too many public methods' pylint
warning, as it crosses the 50 methods limit.

Signed-off-by: Manuel Franceschini <livewire@google.com>
Reviewed-by: Iustin Pop <iustin@google.com>