Make Hostname object always intialize its name to fqdn
This patch restores the behaviour of Hostname (previously known as
HostInfo) to always use fqdn. This was broken due to the fact that the
now used getaddrinfo does not return an fqdn in contrast to
gethostbyname_ex().

This patch adds Hostname.GetFqdn() which is always used when
initializing a Hostname object.

A problem this solves is when adding a node and not using its fqdn, node
verify won't complain.

Signed-off-by: Manuel Franceschini <livewire@google.com>
Reviewed-by: Michael Hanselmann <hansmi@google.com>